Move skin background to separate file
